fix gcnn to work with TriangularB
